Made MvvmNanoMasterDetailPage better customizable and provided a default implementation using ListView.
