+++ src/MvvmNano.Forms/MvvmNano.Forms.csproj
@@ -34,9 +34,10 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="Internals\ViewViewModelHelper.cs" />
-    <Compile Include="MvvmNanoMasterDetailData.cs" />
-    <Compile Include="MvvmNanoMasterDetailPage.cs" />
-    <Compile Include="MvvmNanoMasterDetailPageBase.cs" />
+    <Compile Include="MasterDetail\MvvmNanoDefaultMasterDetailPage.cs" />
+    <Compile Include="MasterDetail\MvvmNanoMasterDetailData.cs" />
+    <Compile Include="MasterDetail\MvvmNanoMasterDetailPage.cs" />
+    <Compile Include="MasterDetail\MvvmNanoMasterDetailPageBase.cs" />
     <Compile Include="Properties\Annotations.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="MvvmNanoContentPage.cs" />
